requirements.txt
Django==3.1.4
psycopg2-binary
django-grappelli
djangorestframework
markdown      
django-filter 